backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pdfplumber==0.11.0
python-multipart==0.0.6
google-generativeai>=0.8.0
sqlalchemy>=2.0.36
